requirements.txt
flask
PyPDF2
docx2txt
scikit-learn
gunicorn
python-dotenv